requirements.txt
absl-py==2.4.0
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
basicsr==1.4.2
brisque==0.0.17
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
facexlib==0.3.0
fastapi==0.128.0
filelock==3.20.3
filterpy==1.4.5
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.76.0
h11==0.16.0
httptools==0.7.1
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
libsvm-official==3.37.0
llvmlite==0.46.0
lmdb==1.7.5
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
realesrgan==0.3.0
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2026.1.28
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
yapf==0.43.0
